First attempt at fixing listener problem.
element.on() register to the events in the bubbling phase. Registering on the document root is a bad idea because any element could actually stop the propagation of the event and then the functionality is broken. Since I dont't see how jqlite can be made to register the event listener in the capturing phase, I am doing this by hand here.